requirements/base.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==4.0.4
django-filter==23.5
django-oauth-toolkit==2.3.0
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
jwcrypto==1.5.6
Markdown==3.6
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
pip==24.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
setuptools==46.1.3
sqlparse==0.4.4
typing_extensions==4.11.0
urllib3==2.2.1
wheel==0.34.2
zope.event==5.0
zope.interface==6.3
ipython==8.23.0
django-extensions==3.2.3
django-hashid-field==3.4.0
pytz==2024.1
django-cors-headers==4.3.1
sentry-sdk==1.45.0